+++ FNA/src/Graphics/Effect/EffectParameter.cs
@@ -1,6 +1,6 @@
 #region License
 /* FNA - XNA4 Reimplementation for Desktop Platforms
- * Copyright 2009-2022 Ethan Lee and the MonoGame Team
+ * Copyright 2009-2023 Ethan Lee and the MonoGame Team
  *
  * Released under the Microsoft Public License.
  * See LICENSE for details.
@@ -538,7 +538,7 @@ namespace Microsoft.Xna.Framework.Graphics
 				unsafe
 				{
 					float *dstPtr = (float*) values;
-					*dstPtr = value;
+					*dstPtr = (float) value;
 				}
 			}
 			else
